Misc py3 porting fixes
I only changed the str(Exception) where I encountered it as an issue,
but it's probably a hint that the other places it's done should also be
changed.

Most of this is a fallback case for if EDID wasn't successfully
extracted, so it's probably rarely used. And I'm not convinced the
fallback will ever work anyway, or at least "xrandr-DisplayPort-0"
didn't match against any outputs on my system...