fix setuptools module in DisplayCAL/setup.py
For some reason "setuptools" is used as a variable and a module at the
same time.

Initially this leads to "setuptools.findall" not being recognized as
a command related to the module and failing as a result.

Any use of module name in conditional statements also leads to errors,
hence the simple solution: apply use_setuptools for this case.