fix version generation in setup.py
setup.py was tied to generating DisplayCAL/__version__.py provided the
.git directory is present and git is found in the environment.

It poses problems in packaging for various (i.e. linux) distributions
as well:

1. Any sane distribution uses release tarballs which do not (and should
not) contain the .git directory.

2. Tarballs already include VERSION_BASE containing release version of
DisplayCAL, rendering git pointless.

With VERSION_BASE present, the only use left for git is to fetch date
of the last commit in the repository.

This may be useful for developers to run various snapshots, but for
distribution purpose it makes no sense - every version has its own
release date (e.g. posted in CHANGES.html) which is quite enough for
packagers and users.